Добавив дисципліну та можливість її добавити видалити
Все окремо працює будемо думати як поєднати дисципліни з оцінками

+++ src/main/java/com/edu/webappstuddb_v4/models/Grade.java
@@ -16,9 +16,9 @@ public class Grade {
     @Column(name = "id")
     private Long id;
     @Column(name = "studentId")
-    private int studentId;
+    private Long studentId;
     @Column(name = "disciplineId")
-    private int disciplineId;
+    private Long disciplineId;
     @Column(name = "subject")
     private String subject;
     @Column(name = "grade")
